fix some errors

+++ src/lib/Device.ts
@@ -12,27 +12,31 @@ export default class Device{
     constructor(private userId:string, private device:string, private configClass:Config, private mqtt:Mqtt){
         console.log("new device", device);
         this.config = configClass.get("device", this.userId, this.device);
-        configClass.on("device", (userId, device, config)=>{
+        configClass.on("device", async (userId, device, config)=>{
             if (userId !== this.userId || device !== this.device) return;
-            this.onConfigChanged(config);
+            await this.onConfigChanged(config);
         });
     }
     async open(){
-        this.switch = new Switch(this.mqtt, this.config.topic);
+        this.switch = new Switch(this.mqtt, this.userId, this.config.topic, this.configClass);
         await this.switch.open();
         this.mode = await modeFactory(this.config, this.switch, this.configClass, this.userId, this.device);
     }
     async delete(){
         console.log("delete device", this.device);
-        await this.mode.delete();
-        await this.switch.delete();
+        if (this.mode) await this.mode.delete();
+        if (this.switch) await this.switch.delete();
     }
     async close(){
         console.log("close device", this.device);
         await this.mode.close();
         await this.switch.close();
     }
-    private onConfigChanged(config:DeviceConfig){
-
+    private async onConfigChanged(config:DeviceConfig){
+        if (this.config.mode !== config.mode){
+            this.config = config;
+            if (this.mode) await this.mode.delete;
+            this.mode = await modeFactory(this.config, this.switch,this.configClass, this.userId, this.device);
+        }
     }
 }
